davinci: gps: Convert enum types to int32_t for comparison

+++ gps/gnss/GnssAdapter.cpp
@@ -4086,7 +4086,7 @@ bool GnssAdapter::needReportForClient(LocationAPI* client, enum loc_sess_status 
         // report intermediate fix when TBT session allows, like flp;
         // report any fix (even failed fix) when TBT session allows, like LE.
         if ((it->first.client == client || client == nullptr) &&
-                it->second.qualityLevelAccepted >= status) {
+                (int32_t ) (it->second.qualityLevelAccepted) >= (int32_t) status) {
             return true;
         }
     }
